requirements.txt
gunicorn
flask
flask_wtf
wtforms
wtforms[email]
flask-sqlalchemy
flask-login
flask-migrate
flask-mysqldb
pyjwt
